Move implementation of commands to a seperate class
